+++ src/internal/services/notifications/controllers.go
@@ -1,0 +1,7 @@
+package service
+
+// Controllers - описание контроллеров сервиса.
+type Controllers interface {}
+
+// controllers - контроллеры сервиса.
+type controllers struct {}