+++ src/internal/services/url_shortner/infrastructure/adapters/urls_management/grpc.go
@@ -3,12 +3,11 @@ package urls_management_adapter
 import (
 	"context"
 	urls_management_controller "sm-box/internal/services/url_shortner/infrastructure/controllers/urls_management"
+	"sm-box/internal/services/url_shortner/objects/constructors"
 	"sm-box/internal/services/url_shortner/objects/models"
-	"sm-box/internal/services/url_shortner/objects/types"
 	"sm-box/pkg/core/components/logger"
 	"sm-box/pkg/core/components/tracer"
 	c_errors "sm-box/pkg/errors"
-	"time"
 )
 
 const (
@@ -20,12 +19,7 @@ type Adapter_Grpc struct {
 	components *components
 
 	controller interface {
-		Create(ctx context.Context,
-			source string,
-			type_ types.ShortUrlType,
-			numberOfUses int64,
-			startActive, endActive time.Time,
-		) (url *models.ShortUrlInfo, cErr c_errors.Error)
+		Create(ctx context.Context, constructor *constructors.ShortUrl) (url *models.ShortUrlInfo, cErr c_errors.Error)
 	}
 
 	ctx context.Context
@@ -71,22 +65,18 @@ func New_Grpc(ctx context.Context) (adapter *Adapter_Grpc, err error) {
 }
 
 // Create - создание сокращенного url.
-func (adapter *Adapter_Grpc) Create(ctx context.Context,
-	source string,
-	type_ types.ShortUrlType,
-	numberOfUses int64,
-	startActive, endActive time.Time) (url *models.ShortUrlInfo, cErr c_errors.Grpc) {
+func (adapter *Adapter_Grpc) Create(ctx context.Context, constructor *constructors.ShortUrl) (url *models.ShortUrlInfo, cErr c_errors.Grpc) {
 	// tracer
 	{
 		var trc = tracer.New(tracer.LevelAdapter)
 
-		trc.FunctionCall(ctx, source, type_, numberOfUses, startActive, endActive)
+		trc.FunctionCall(ctx, constructor)
 		defer func() { trc.Error(cErr).FunctionCallFinished(url) }()
 	}
 
 	var proxyErr c_errors.Error
 
-	if url, proxyErr = adapter.controller.Create(ctx, source, type_, numberOfUses, startActive, endActive); proxyErr != nil {
+	if url, proxyErr = adapter.controller.Create(ctx, constructor); proxyErr != nil {
 		cErr = c_errors.ToGrpc(proxyErr)
 
 		adapter.components.Logger.Error().