+++ src/migrations/notifications.sql
@@ -1,0 +1,4 @@
+create database notifications;
+
+create schema
+    if not exists public;